requirements.txt
moviepy
google-auth
google-api-python-client
google-auth-oauthlib
requests
pytz
schedule
python-dotenv
google-genai
google-cloud-texttospeech
pydub
nltk
audioop-lts